requirements.txt
streamlit
numpy
tensorflow
keras
Pillow
